Refactor GitHub Actions workflow for building and testing
Simplify the GitHub Actions workflow for building and testing the project.
The previous workflow file (.github/workflows/test.yml) was deleted, and a new file (.github/workflows/build-test.yml) was added.
The new workflow file includes updated versions of the Docker actions and removes unused files.
The Go version in the release-binary.yml file was also updated to 1.23.
Additionally, the package name in web/package-lock.json was fixed.